bug: cannot remove sync folder. Improve sync management

+++ omyc/www/sync/views/view/view.js
@@ -57,6 +57,7 @@ app.controller('view', function($scope,$routeParams,$location,$http,$interval) {
             $scope.api.response = { error:{ code:"ID_EMPTY", message:"Id is empty"} };
             return
         }
+        if (!confirm("Confirm delete?")) { return }
         //
         $scope.api.busy 	= true;
         $http.delete($scope.api.url+$scope.item._id)
